move init stuff to init.go
